+++ packages/theme/docs/.vitepress/blog-theme.ts
@@ -69,8 +69,7 @@ export const blogTheme = getThemeConfig({
     ]
   },
   oml2d: {
-    mobileDisplay: true,
-    models: [
+    model: [
       {
         path: 'https://oml2d-models.sugarat.top/Senko_Normals/senko.model3.json',
       },