+++ packages/theme/src/components/BlogOml2d.vue
@@ -1,10 +1,10 @@
 <script lang="ts" setup>
 import { useOml2d } from '../hooks/useOml2d'
 
-// oh-my-live2d 扩展
+// l2d-widget 扩展
 useOml2d()
 </script>
 
 <template>
-  <div id="oh-my-live2d" style="display: none;" />
+  <div id="l2d-widget" style="display: none;" />
 </template>